feat(nvim): Add lazy require function
This commit introduces a new function, `lazy_require`, to the Neovim
configuration. This function allows for modules to be required lazily,
which can improve startup time and performance. The function is a
wrapper around the `require` function, returning a metatable that
defers the actual require until a function from the module is called.

See: https://github.com/tjdevries/lazy-require.nvim/blob/master/lua/lazy-require.lua